requirements.txt
Django==1.5.1
djangorestframework==2.3.5